requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
